// ui/src/lib/api/model/internalMonitoringQueueStatsExtended.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * BrowserGrid API
 * BrowserGrid is a distributed browser automation platform using task queues for scalable browser session management.
 * OpenAPI spec version: 2.0
 */

export interface InternalMonitoringQueueStatsExtended {
  active?: number;
  archived?: number;
  error_rate?: number;
  failed?: number;
  last_activity?: string;
  memory_usage?: string;
  pending?: number;
  processed?: number;
  queue?: string;
  retry?: number;
  scheduled?: number;
  size?: number;
  /** "run", "paused" */
  state?: string;
}